+++ src/Sentry/Internal/SdkComposer.cs
@@ -1,5 +1,5 @@
 using System;
-using System.Threading;
+using System.Threading.Tasks;
 using Sentry.Extensibility;
 using Sentry.Internal.Http;
 
@@ -53,23 +53,34 @@ namespace Sentry.Internal
                     _options.InitCacheFlushTimeout
                 );
 
-                // Use a timeout to avoid waiting for too long
-                using var timeout = new CancellationTokenSource(_options.InitCacheFlushTimeout);
-
                 try
                 {
-                    cachingTransport.FlushAsync(timeout.Token).GetAwaiter().GetResult();
-                }
-                catch (OperationCanceledException)
-                {
-                    _options.DiagnosticLogger?.LogError(
-                        "Flushing timed out."
-                    );
+                    // Flush cache but block on it only for a limited amount of time.
+                    // If we don't flush it in time, then continue doing it on the
+                    // background but don't block the calling thread until it finishes.
+                    var timeoutTask = Task.Delay(_options.InitCacheFlushTimeout);
+                    var flushTask = cachingTransport.FlushAsync();
+
+                    // If flush finished in time, finalize the task by awaiting it to
+                    // propagate potential exceptions.
+                    if (Task.WhenAny(timeoutTask, flushTask).GetAwaiter().GetResult() == flushTask)
+                    {
+                        flushTask.GetAwaiter().GetResult();
+                    }
+                    // If flush timed out, log and continue
+                    else
+                    {
+                        _options.DiagnosticLogger?.LogInfo(
+                            "Cache flushing is taking longer than the configured timeout of {0}. " +
+                            "Continuing without waiting for the task to finish.",
+                            _options.InitCacheFlushTimeout
+                        );
+                    }
                 }
                 catch (Exception ex)
                 {
-                    _options.DiagnosticLogger?.LogFatal(
-                        "Flushing failed.",
+                    _options.DiagnosticLogger?.LogError(
+                        "Cache flushing failed.",
                         ex
                     );
                 }